Replaced doneRenderingComponent event with an observable and removed componentDoneUnloading event. #2632
