+++ src/main/webapp/wise5/components/animation/animationAuthoringController.ts
@@ -9,6 +9,7 @@ class AnimationAuthoringController extends AnimationController {
 
   static $inject = [
     '$filter',
+    '$injector',
     '$mdDialog',
     '$q',
     '$rootScope',
@@ -28,6 +29,7 @@ class AnimationAuthoringController extends AnimationController {
 
   constructor(
     $filter,
+    $injector,
     $mdDialog,
     $q,
     $rootScope,
@@ -46,6 +48,7 @@ class AnimationAuthoringController extends AnimationController {
   ) {
     super(
       $filter,
+      $injector,
       $mdDialog,
       $q,
       $rootScope,
