+++ src/main/webapp/wise5/components/animation/animationController.ts
@@ -28,6 +28,7 @@ class AnimationController extends ComponentController {
 
   static $inject = [
     '$filter',
+    '$injector',
     '$mdDialog',
     '$q',
     '$rootScope',
@@ -46,6 +47,7 @@ class AnimationController extends ComponentController {
 
   constructor(
     $filter,
+    $injector,
     $mdDialog,
     $q,
     $rootScope,
@@ -63,6 +65,7 @@ class AnimationController extends ComponentController {
   ) {
     super(
       $filter,
+      $injector,
       $mdDialog,
       $q,
       $rootScope,
