+++ src/main/webapp/wise5/components/audioOscillator/audioOscillatorAuthoringController.ts
@@ -13,6 +13,7 @@ class AudioOscillatorAuthoringController extends AudioOscillatorController {
 
   static $inject = [
     '$filter',
+    '$injector',
     '$mdDialog',
     '$q',
     '$rootScope',
@@ -32,6 +33,7 @@ class AudioOscillatorAuthoringController extends AudioOscillatorController {
 
   constructor(
     $filter,
+    $injector,
     $mdDialog,
     $q,
     $rootScope,
@@ -50,6 +52,7 @@ class AudioOscillatorAuthoringController extends AudioOscillatorController {
   ) {
     super(
       $filter,
+      $injector,
       $mdDialog,
       $q,
       $rootScope,
