+++ src/main/webapp/wise5/components/audioOscillator/audioOscillatorController.ts
@@ -31,6 +31,7 @@ class AudioOscillatorController extends ComponentController {
 
   static $inject = [
     '$filter',
+    '$injector',
     '$mdDialog',
     '$q',
     '$rootScope',
@@ -49,6 +50,7 @@ class AudioOscillatorController extends ComponentController {
 
   constructor(
     $filter,
+    $injector,
     $mdDialog,
     $q,
     $rootScope,
@@ -66,6 +68,7 @@ class AudioOscillatorController extends ComponentController {
   ) {
     super(
       $filter,
+      $injector,
       $mdDialog,
       $q,
       $rootScope,
