+++ src/main/webapp/wise5/components/conceptMap/conceptMapAuthoringController.ts
@@ -17,6 +17,7 @@ class ConceptMapAuthoringController extends ConceptMapController {
   static $inject = [
     '$anchorScroll',
     '$filter',
+    '$injector',
     '$location',
     '$mdDialog',
     '$q',
@@ -38,6 +39,7 @@ class ConceptMapAuthoringController extends ConceptMapController {
   constructor(
     $anchorScroll,
     $filter,
+    $injector,
     $location,
     $mdDialog,
     $q,
@@ -58,6 +60,7 @@ class ConceptMapAuthoringController extends ConceptMapController {
     super(
       $anchorScroll,
       $filter,
+      $injector,
       $location,
       $mdDialog,
       $q,
