+++ src/main/webapp/wise5/components/discussion/discussionAuthoringController.ts
@@ -9,6 +9,7 @@ class DiscussionAuthoringController extends DiscussionController {
 
   static $inject = [
     '$filter',
+    '$injector',
     '$mdDialog',
     '$q',
     '$rootScope',
@@ -29,6 +30,7 @@ class DiscussionAuthoringController extends DiscussionController {
 
   constructor(
     $filter,
+    $injector,
     $mdDialog,
     $q,
     $rootScope,
@@ -48,6 +50,7 @@ class DiscussionAuthoringController extends DiscussionController {
   ) {
     super(
       $filter,
+      $injector,
       $mdDialog,
       $q,
       $rootScope,
