+++ src/main/webapp/wise5/components/embedded/embeddedAuthoringController.ts
@@ -10,6 +10,7 @@ class EmbeddedAuthoringController extends EmbeddedController {
 
   static $inject = [
     '$filter',
+    '$injector',
     '$mdDialog',
     '$q',
     '$rootScope',
@@ -30,6 +31,7 @@ class EmbeddedAuthoringController extends EmbeddedController {
   ];
 
   constructor($filter,
+              $injector,
               $mdDialog,
               $q,
               $rootScope,
@@ -48,6 +50,7 @@ class EmbeddedAuthoringController extends EmbeddedController {
               StudentDataService,
               UtilService) {
     super($filter,
+      $injector,
       $mdDialog,
       $q,
       $rootScope,
