+++ src/main/webapp/wise5/components/embedded/embeddedController.ts
@@ -25,6 +25,7 @@ class EmbeddedController extends ComponentController {
 
   static $inject = [
     '$filter',
+    '$injector',
     '$mdDialog',
     '$q',
     '$rootScope',
@@ -45,6 +46,7 @@ class EmbeddedController extends ComponentController {
 
   constructor(
     $filter,
+    $injector,
     $mdDialog,
     $q,
     $rootScope,
@@ -64,6 +66,7 @@ class EmbeddedController extends ComponentController {
   ) {
     super(
       $filter,
+      $injector,
       $mdDialog,
       $q,
       $rootScope,
