+++ src/main/webapp/wise5/components/graph/graphAuthoringController.ts
@@ -19,6 +19,7 @@ class GraphAuthoringController extends GraphController {
 
   static $inject = [
     '$filter',
+    '$injector',
     '$mdDialog',
     '$q',
     '$rootScope',
@@ -37,6 +38,7 @@ class GraphAuthoringController extends GraphController {
   ];
 
   constructor($filter,
+              $injector,
               $mdDialog,
               $q,
               $rootScope,
@@ -53,6 +55,7 @@ class GraphAuthoringController extends GraphController {
               StudentDataService,
               UtilService) {
     super($filter,
+      $injector,
       $mdDialog,
       $q,
       $rootScope,
