+++ src/main/webapp/wise5/components/graph/graphController.ts
@@ -63,6 +63,7 @@ class GraphController extends ComponentController {
 
   static $inject = [
     '$filter',
+    '$injector',
     '$mdDialog',
     '$q',
     '$rootScope',
@@ -81,6 +82,7 @@ class GraphController extends ComponentController {
 
   constructor(
     $filter,
+    $injector,
     $mdDialog,
     $q,
     $rootScope,
@@ -98,6 +100,7 @@ class GraphController extends ComponentController {
   ) {
     super(
       $filter,
+      $injector,
       $mdDialog,
       $q,
       $rootScope,
@@ -3034,7 +3037,7 @@ class GraphController extends ComponentController {
    * @return A promise that returns the url of the image that is generated from the component state.
    */
   setComponentStateAsBackgroundImage(componentState) {
-    return this.UtilService.generateImageFromComponentState(componentState).then(image => {
+    return this.generateImageFromComponentState(componentState).then(image => {
       return image.url;
     });
   }
