+++ src/main/webapp/wise5/components/html/htmlAuthoringController.ts
@@ -11,6 +11,7 @@ class HTMLAuthoringController extends HTMLController {
   summernotePromptId: string;
 
   static $inject = [
+    '$injector',
     '$q',
     '$rootScope',
     '$scope',
@@ -31,6 +32,7 @@ class HTMLAuthoringController extends HTMLController {
   ];
 
   constructor(
+    $injector,
     $q,
     $rootScope,
     $scope,
@@ -50,6 +52,7 @@ class HTMLAuthoringController extends HTMLController {
     UtilService
   ) {
     super(
+      $injector,
       $q,
       $rootScope,
       $scope,
