+++ src/main/webapp/wise5/components/html/htmlController.ts
@@ -9,6 +9,7 @@ class HTMLController extends ComponentController {
   html: string;
 
   static $inject = [
+    '$injector',
     '$q',
     '$rootScope',
     '$scope',
@@ -28,6 +29,7 @@ class HTMLController extends ComponentController {
   ];
 
   constructor(
+    $injector,
     $q,
     $rootScope,
     $scope,
@@ -47,6 +49,7 @@ class HTMLController extends ComponentController {
   ) {
     super(
       $filter,
+      $injector,
       $mdDialog,
       $q,
       $rootScope,
@@ -72,10 +75,7 @@ class HTMLController extends ComponentController {
       }
     }
 
-    this.$rootScope.$broadcast('doneRenderingComponent', {
-      nodeId: this.nodeId,
-      componentId: this.componentId
-    });
+    this.broadcastDoneRenderingComponent();
   }
 }
