+++ src/main/webapp/wise5/components/label/labelAuthoringController.ts
@@ -14,6 +14,7 @@ class LabelAuthoringController extends LabelController {
 
   static $inject = [
     '$filter',
+    '$injector',
     '$mdDialog',
     '$q',
     '$rootScope',
@@ -34,6 +35,7 @@ class LabelAuthoringController extends LabelController {
 
   constructor(
     $filter,
+    $injector,
     $mdDialog,
     $q,
     $rootScope,
@@ -53,6 +55,7 @@ class LabelAuthoringController extends LabelController {
   ) {
     super(
       $filter,
+      $injector,
       $mdDialog,
       $q,
       $rootScope,
