+++ src/main/webapp/wise5/components/match/matchAuthoringController.ts
@@ -10,6 +10,7 @@ class MatchAuthoringController extends MatchController {
 
   static $inject = [
     '$filter',
+    '$injector',
     '$mdDialog',
     '$mdMedia',
     '$q',
@@ -30,6 +31,7 @@ class MatchAuthoringController extends MatchController {
 
   constructor(
     $filter,
+    $injector,
     $mdDialog,
     $mdMedia,
     $q,
@@ -49,6 +51,7 @@ class MatchAuthoringController extends MatchController {
   ) {
     super(
       $filter,
+      $injector,
       $mdDialog,
       $mdMedia,
       $q,
