+++ src/main/webapp/wise5/components/match/matchController.ts
@@ -26,6 +26,7 @@ class MatchController extends ComponentController {
 
   static $inject = [
     '$filter',
+    '$injector',
     '$mdDialog',
     '$mdMedia',
     '$q',
@@ -45,6 +46,7 @@ class MatchController extends ComponentController {
 
   constructor(
     $filter,
+    $injector,
     $mdDialog,
     $mdMedia,
     $q,
@@ -63,6 +65,7 @@ class MatchController extends ComponentController {
   ) {
     super(
       $filter,
+      $injector,
       $mdDialog,
       $q,
       $rootScope,
@@ -198,10 +201,7 @@ class MatchController extends ComponentController {
       return deferred.promise;
     };
 
-    this.$rootScope.$broadcast('doneRenderingComponent', {
-      nodeId: this.nodeId,
-      componentId: this.componentId
-    });
+    this.broadcastDoneRenderingComponent();
   }
 
   addNotebookItemToSourceBucket(notebookItem) {
