+++ src/main/webapp/wise5/components/multipleChoice/multipleChoiceAuthoringController.ts
@@ -9,6 +9,7 @@ class MultipleChoiceAuthoringController extends MultipleChoiceController {
 
   static $inject = [
     '$filter',
+    '$injector',
     '$mdDialog',
     '$q',
     '$rootScope',
@@ -27,6 +28,7 @@ class MultipleChoiceAuthoringController extends MultipleChoiceController {
 
   constructor(
     $filter,
+    $injector,
     $mdDialog,
     $q,
     $rootScope,
@@ -44,6 +46,7 @@ class MultipleChoiceAuthoringController extends MultipleChoiceController {
   ) {
     super(
       $filter,
+      $injector,
       $mdDialog,
       $q,
       $rootScope,
