+++ src/main/webapp/wise5/components/multipleChoice/multipleChoiceController.ts
@@ -15,6 +15,7 @@ class MultipleChoiceController extends ComponentController {
 
   static $inject = [
     '$filter',
+    '$injector',
     '$mdDialog',
     '$q',
     '$rootScope',
@@ -32,6 +33,7 @@ class MultipleChoiceController extends ComponentController {
 
   constructor(
     $filter,
+    $injector,
     $mdDialog,
     $q,
     $rootScope,
@@ -48,6 +50,7 @@ class MultipleChoiceController extends ComponentController {
   ) {
     super(
       $filter,
+      $injector,
       $mdDialog,
       $q,
       $rootScope,
@@ -205,10 +208,7 @@ class MultipleChoiceController extends ComponentController {
       angular.bind(this, function(event, args) {})
     );
 
-    this.$rootScope.$broadcast('doneRenderingComponent', {
-      nodeId: this.nodeId,
-      componentId: this.componentId
-    });
+    this.broadcastDoneRenderingComponent();
   }
 
   handleNodeSubmit() {
