+++ src/main/webapp/wise5/components/openResponse/openResponseController.ts
@@ -29,6 +29,7 @@ class OpenResponseController extends ComponentController {
 
   static $inject = [
     '$filter',
+    '$injector',
     '$mdDialog',
     '$q',
     '$rootScope',
@@ -49,6 +50,7 @@ class OpenResponseController extends ComponentController {
 
   constructor(
     $filter,
+    $injector,
     $mdDialog,
     $q,
     $rootScope,
@@ -68,6 +70,7 @@ class OpenResponseController extends ComponentController {
   ) {
     super(
       $filter,
+      $injector,
       $mdDialog,
       $q,
       $rootScope,
@@ -289,10 +292,7 @@ class OpenResponseController extends ComponentController {
       });
     }
 
-    this.$rootScope.$broadcast('doneRenderingComponent', {
-      nodeId: this.nodeId,
-      componentId: this.componentId
-    });
+    this.broadcastDoneRenderingComponent();
   }
 
   handleNodeSubmit() {
