+++ src/main/webapp/wise5/components/outsideURL/outsideURLAuthoringController.ts
@@ -13,6 +13,7 @@ class OutsideURLAuthoringController extends OutsideURLController {
 
   static $inject = [
     '$filter',
+    '$injector',
     '$mdDialog',
     '$q',
     '$rootScope',
@@ -32,6 +33,7 @@ class OutsideURLAuthoringController extends OutsideURLController {
 
   constructor(
     $filter,
+    $injector,
     $mdDialog,
     $q,
     $rootScope,
@@ -50,6 +52,7 @@ class OutsideURLAuthoringController extends OutsideURLController {
   ) {
     super(
       $filter,
+      $injector,
       $mdDialog,
       $q,
       $rootScope,
