+++ src/main/webapp/wise5/components/outsideURL/outsideURLController.ts
@@ -14,6 +14,7 @@ class OutsideURLController extends ComponentController {
 
   static $inject = [
     '$filter',
+    '$injector',
     '$mdDialog',
     '$q',
     '$rootScope',
@@ -32,6 +33,7 @@ class OutsideURLController extends ComponentController {
 
   constructor(
     $filter,
+    $injector,
     $mdDialog,
     $q,
     $rootScope,
@@ -49,6 +51,7 @@ class OutsideURLController extends ComponentController {
   ) {
     super(
       $filter,
+      $injector,
       $mdDialog,
       $q,
       $rootScope,
@@ -75,10 +78,7 @@ class OutsideURLController extends ComponentController {
 
     this.setWidthAndHeight(this.componentContent.width, this.componentContent.height);
 
-    this.$rootScope.$broadcast('doneRenderingComponent', {
-      nodeId: this.nodeId,
-      componentId: this.componentId
-    });
+    this.broadcastDoneRenderingComponent();
   }
 
   setWidthAndHeight(width, height) {
