+++ src/main/webapp/wise5/components/table/tableAuthoringController.ts
@@ -10,6 +10,7 @@ class TableAuthoringController extends TableController {
   constructor(
     $anchorScroll,
     $filter,
+    $injector,
     $location,
     $mdDialog,
     $q,
@@ -29,6 +30,7 @@ class TableAuthoringController extends TableController {
     super(
       $anchorScroll,
       $filter,
+      $injector,
       $location,
       $mdDialog,
       $q,
@@ -577,6 +579,7 @@ class TableAuthoringController extends TableController {
 TableAuthoringController.$inject = [
   '$anchorScroll',
   '$filter',
+  '$injector',
   '$location',
   '$mdDialog',
   '$q',
