+++ src/main/webapp/wise5/services/nodeService.ts
@@ -7,6 +7,7 @@ import { ProjectService } from './projectService';
 import { UpgradeModule } from '@angular/upgrade/static';
 import { ChooseBranchPathDialogComponent } from '../../site/src/app/preview/modules/choose-branch-path-dialog/choose-branch-path-dialog.component';
 import { DataService } from '../../site/src/app/services/data.service';
+import { Subject } from 'rxjs';
 
 @Injectable()
 export class NodeService {
@@ -14,6 +15,8 @@ export class NodeService {
   $translate: any;
   transitionResults = {};
   chooseTransitionPromises = {};
+  private doneRenderingComponentSource: Subject<any> = new Subject<any>();
+  public doneRenderingComponent$ = this.doneRenderingComponentSource.asObservable();
 
   constructor(
     private upgrade: UpgradeModule,
@@ -785,5 +788,9 @@ export class NodeService {
       clickOutsideToClose: true,
       escapeToClose: true
     });
+  }
+
+  broadcastDoneRenderingComponent(nodeIdAndComponentId: any) {
+    this.doneRenderingComponentSource.next(nodeIdAndComponentId);
   }
 }
