+++ src/main/webapp/wise5/themes/default/notebook/notebook/notebook.js
@@ -90,7 +90,6 @@ class NotebookController {
 
     this.logOutListener = $scope.$on('logOut', (event, args) => {
       this.logOutListener();
-      this.$rootScope.$broadcast('componentDoneUnloading');
     });
 
     this.notebook = this.NotebookService.getNotebookByWorkgroup(this.workgroupId);
